Task 4 done

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -11,7 +11,7 @@
 
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int shift_left, invert;
+	unsigned long int shift_left, invert;
 
 	if (index >= sizeof(unsigned long int) * 8)
 		return (-1);
